pyproject.toml
[tool.poetry]
name = "seq2ids"
version = "0.1.0"
description = "blast sequences to get ids of the taxon. gene, etc."
authors = ["Mark Andrew Miller <MAM@lbl.gov>"]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.79"
click = "^8.0.4"
click-log = "^0.3.2"
SQLAlchemy = "^1.4.32"
#psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.1"
PyYAML = "^6.0"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
bcbio-gff = "^0.6.9"
gffutils = "^0.10.1"
black = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
seq2ids = "seq2ids.seq2ids:seq2ids"
